+++ src/components/nav/components/seach_bar/hooks.tsx
@@ -1,3 +1,4 @@
+import numeral from 'numeral';
 import { chainConfig } from '@configs';
 import { useRouter } from 'next/router';
 import {
@@ -33,8 +34,8 @@ export const useSearchBar = (t) => {
         router.push(VALIDATOR_DETAILS(value));
       } else if (new RegExp(userRegex).test(value)) {
         router.push(ACCOUNT_DETAILS(value));
-      } else if (/^-?\d+$/.test(value)) {
-        router.push(BLOCK_DETAILS(value));
+      } else if (/^-?\d+$/.test(numeral(value).value())) {
+        router.push(BLOCK_DETAILS(numeral(value).value()));
       } else {
         router.push(TRANSACTION_DETAILS(value));
       }